Use 32/64 independent int types in pak_util
The previous version would work on PAK-files as long as they were written
and read on the same platform. That meant that a file written on a 32-bit
platform couldn't be read on a 64-bit platform. Using this patch, it should
work as intended.

+++ source/pak_util.h
@@ -44,6 +44,7 @@
   extern "C" {
 #endif
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -56,14 +57,14 @@
 
 typedef struct{
   char name[56];
-  unsigned long int offset;
-  unsigned long int size;
+  uint32_t offset;
+  uint32_t size;
 }PAK_DirEntry;
 
 typedef struct{
   FILE *PAK_archive;
-  unsigned long int offset;
-  unsigned long int size;
+  uint32_t offset;
+  uint32_t size;
   long filepos;
 }PAK_file;
